# contracts/project-control-v1/openapi.yaml
openapi: 3.1.0
info:
  title: Zhishu Project Control Internal API
  version: 1.19.0
  description: >-
    Service-credential protected Node BFF to Control Plane protocol. Execution reads
    reuse existing projections; semantic state writes and parallel scheduling are
    proposal-first and actor-audited. Workspace provisioning is a controlled
    marked-directory operation; Integration evidence is captured as immutable snapshots,
    and Git-backed Worktree operations are not implicit. Successful Integration candidates
    require an explicit ownership-checked, clean-primary fast-forward finalization request.
servers:
  - url: http://127.0.0.1:8080
security:
  - projectControlToken: []
paths:
...
  /internal/project-control/v1/projects/{projectId}/work-orders:
    get:
      operationId: listWorkOrders
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - { name: planId, in: query, required: false, schema: { type: string, format: uuid } }
        - { name: nodeKey, in: query, required: false, schema: { type: string, minLength: 1 } }
      responses:
        '200':
          description: Historical Work Orders and their existing Task bindings.
          content:
            application/json:
              schema: { type: array, items: { $ref: '../work-order-v1/work-order.schema.json' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /internal/project-control/v1/projects/{projectId}/work-orders/{workOrderId}:
    get:
      operationId: getWorkOrder
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - { $ref: '#/components/parameters/WorkOrderId' }
      responses:
        '200':
          description: Replayable Work Order, binding, reports, verifications and events.
          content:
            application/json:
              schema: { $ref: '../work-order-v1/work-order.schema.json' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '../control-v1/openapi.yaml#/components/responses/Conflict' }
  /internal/project-control/v1/projects/{projectId}/work-orders/claims:
    post:
      operationId: claimAndDispatchWorkOrder
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - { $ref: '#/components/parameters/Actor' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClaimWorkOrderRequest' }
      responses:
        '201':
          description: Idempotently claimed Work Order bound to exactly one existing Task and first Attempt.
          content:
            application/json:
              schema: { $ref: '../work-order-v1/work-order.schema.json' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '../control-v1/openapi.yaml#/components/responses/Conflict' }
  /internal/project-control/v1/projects/{projectId}/work-orders/{workOrderId}/verifications:
    post:
      operationId: verifyExecutionReport
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - { $ref: '#/components/parameters/WorkOrderId' }
        - { $ref: '#/components/parameters/Actor' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VerifyExecutionReportRequest' }
      responses:
        '201': { description: Human Verification plus fresh Brain Context handoff. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '../control-v1/openapi.yaml#/components/responses/Conflict' }
  /internal/project-control/v1/projects/{projectId}/work-orders/{workOrderId}/brain-handoff:
    post:
      operationId: recordWorkOrderBrainHandoff
      parameters:
        - { $ref: '#/components/parameters/ProjectId' }
        - { $ref: '#/components/parameters/WorkOrderId' }
        - { $ref: '#/components/parameters/Actor' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderBrainHandoffRequest' }
      responses:
        '201': { description: Replay-safe audited Brain summary pinned to the post-reducer Context Package. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '../control-v1/openapi.yaml#/components/responses/Conflict' }
components:
  securitySchemes:
    projectControlToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-service-token
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Actor:
      name: X-Zhishu-Actor
      in: header
      required: true
      schema: { type: string, minLength: 1 }
...
    WorkOrderId:
      name: workOrderId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    ClaimWorkOrderRequest:
      type: object
      additionalProperties: false
      required: [clientRequestId, planId, nodeKey, profileId]
      properties:
        clientRequestId: { type: string, minLength: 1, maxLength: 160 }
        planId: { type: string, format: uuid }
        nodeKey: { type: string, minLength: 1 }
        profileId: { type: string, format: uuid }
        approvalMode: { type: string, enum: [MANUAL, AUTO_SAFE, AUTO_TRUSTED] }
    VerifyExecutionReportRequest:
      type: object
      additionalProperties: false
      required: [reportId, decision, reason]
      properties:
        reportId: { type: string, format: uuid }
        decision: { type: string, enum: [PASSED, FAILED] }
        reason: { type: string, minLength: 1 }
    WorkOrderBrainHandoffRequest:
      type: object
      additionalProperties: false
      required:
        - verificationId
        - threadId
        - contextPackageId
        - provider
        - model
        - permissions
        - compositeContextVersion
        - inputTokens
        - outputTokens
        - userMessage
        - brainMessage
        - citations
      properties:
        verificationId: { type: string, format: uuid }
        threadId: { type: string, format: uuid }
        contextPackageId: { type: string, format: uuid }
        provider: { type: string, minLength: 1 }
        model: { type: string, minLength: 1 }
        permissions: { type: object }
        compositeContextVersion: { $ref: '../project-state-v1/project-state.schema.json#/$defs/CompositeContextVersionV1' }
        inputTokens: { type: integer, minimum: 0 }
        outputTokens: { type: integer, minimum: 0 }
        userMessage: { type: string, minLength: 1 }
        brainMessage: { type: string, minLength: 1 }
        citations: { type: array, items: { type: object } }
        generatedProposal: { type: [object, 'null'] }
...
  responses:
    Unauthorized:
      description: Missing or invalid Project Control service credential.
...